+++ packages/tooling/fast-tooling/src/message-system/data.ts
@@ -1,4 +1,4 @@
-import { uniqueId } from "lodash-es";
+import { cloneDeep, get, uniqueId, unset } from "lodash-es";
 import {
     Data,
     DataDictionary,
@@ -89,4 +89,66 @@ export function getLinkedDataDictionary(
         dataDictionary: resolvedDataDictionary[0],
         dictionaryId: config.dictionaryId,
     };
+}
+
+/**
+ * Gets linked data from a data dictionary
+ * that can be used to add to an existing data dictionary
+ * via the Message System
+ */
+export function getLinkedData(
+    dataDictionary: DataDictionary<unknown>,
+    dictionaryIds: string[]
+): Data<unknown>[] {
+    return dictionaryIds.map(dictionaryId => {
+        const data = cloneDeep(dataDictionary[0][dictionaryId].data);
+        // Retrieve all direct linked data for each data item
+        // and removes it from the original parent linked data item
+        const linkedDataKeys: string[] = Object.keys(dataDictionary[0]).filter(
+            (dataDictionaryKey: string) => {
+                const isLinkedData =
+                    get(dataDictionary[0][dataDictionaryKey], "parent.id") ===
+                    dictionaryId;
+
+                if (isLinkedData) {
+                    unset(
+                        data,
+                        get(dataDictionary[0][dataDictionaryKey], "parent.dataLocation")
+                    );
+                }
+
+                return isLinkedData;
+            }
+        );
+
+        return {
+            schemaId: dataDictionary[0][dictionaryId].schemaId,
+            data,
+            linkedData: getLinkedData(dataDictionary, linkedDataKeys),
+        };
+    });
+}
+
+/**
+ * Gets a list of linked data ids from a single dictionary id
+ */
+export function getLinkedDataList(
+    dataDictionary: DataDictionary<unknown>,
+    dictionaryId: string
+): string[] {
+    let linkedDataIds: string[] = [];
+
+    Object.entries(dataDictionary[0]).forEach(
+        ([key, dataDictionaryItem]: [string, Data<unknown>]) => {
+            if (get(dataDictionaryItem, "parent.id") === dictionaryId) {
+                linkedDataIds.push(key);
+
+                linkedDataIds = linkedDataIds.concat(
+                    getLinkedDataList(dataDictionary, key)
+                );
+            }
+        }
+    );
+
+    return linkedDataIds;
 }