+++ packages/web-components/fast-components/src/button/index.ts
@@ -60,3 +60,9 @@ export class FASTButton extends Button {
         }
     }
 }
+
+/**
+ * Styles for Button
+ * @public
+ */
+export const ButtonStyles = styles;