+++ packages/web-components/fast-components/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.7.0](https://github.com/Microsoft/fast/compare/@microsoft/fast-components@1.6.0...@microsoft/fast-components@1.7.0) (2020-07-31)
+
+
+### Bug Fixes
+
+* add form submission and reset behavior to Button ([#3603](https://github.com/Microsoft/fast/issues/3603)) ([f646343](https://github.com/Microsoft/fast/commit/f6463436a001d07a7cfbd5d87460c76cf8584e65))
+* correct class name for text field start/end styles ([#3551](https://github.com/Microsoft/fast/issues/3551)) ([d565dee](https://github.com/Microsoft/fast/commit/d565dee0b8f4f4e25ce4d7ee208a0f3d291d9169))
+* exposing styles on all web components ([#3602](https://github.com/Microsoft/fast/issues/3602)) ([34ea7e2](https://github.com/Microsoft/fast/commit/34ea7e2f3e87a5a27fd9e5475ecab2bc9a44f9eb))
+* fixed error in slider story causing issue in slider-label story ([#3600](https://github.com/Microsoft/fast/issues/3600)) ([ad6e0e9](https://github.com/Microsoft/fast/commit/ad6e0e91571f716854c2c2f1099460e99bc8528f))
+* remove an extra set of quotes in CSS template ([#3594](https://github.com/Microsoft/fast/issues/3594)) ([f70ff99](https://github.com/Microsoft/fast/commit/f70ff9975a801cf8a9863d6a79fc61e62fbd4477))
+* update docusaurus and use local fast-components min bundle ([#3605](https://github.com/Microsoft/fast/issues/3605)) ([db23f27](https://github.com/Microsoft/fast/commit/db23f279dc646eeb043327f2cce7a96f9b89a0c4))
+* update fast-card and component sampler color handling ([#3597](https://github.com/Microsoft/fast/issues/3597)) ([9d32aa1](https://github.com/Microsoft/fast/commit/9d32aa15cd5c18d51c90e35100b0284c5ebb3ca6))
+
+
+### Features
+
+* adding directional stylesheet behavior ([#3559](https://github.com/Microsoft/fast/issues/3559)) ([74c19af](https://github.com/Microsoft/fast/commit/74c19af79cb6b9c015ab3a454a3e69d453f1a217))
+* create an observable function to support aria-valuetext on slider component ([#3508](https://github.com/Microsoft/fast/issues/3508)) ([b4ac369](https://github.com/Microsoft/fast/commit/b4ac369e37ba591a5bb1d8c7bcbfde6dffcadeef)), closes [#3504](https://github.com/Microsoft/fast/issues/3504)
+
+
+
+
+
+# [1.6.0](https://github.com/Microsoft/fast/compare/@microsoft/fast-components@1.5.0...@microsoft/fast-components@1.6.0) (2020-07-23)
+
+
+### Bug Fixes
+
+* address nested styling issues and provide a more intuitive API for nesting ([#3528](https://github.com/Microsoft/fast/issues/3528)) ([4fe0dd3](https://github.com/Microsoft/fast/commit/4fe0dd38ce8f2b43be0f13c7efac2f12ada6cd78))
+* move text field and text area appearances from foundation to components ([#3540](https://github.com/Microsoft/fast/issues/3540)) ([ca8ac76](https://github.com/Microsoft/fast/commit/ca8ac760cdf666fc79a47ba9a21c5f964556dbab))
+* rollup minify selectors should retain spaces ([#3524](https://github.com/Microsoft/fast/issues/3524)) ([cbdfc45](https://github.com/Microsoft/fast/commit/cbdfc45c2543fe9f94e0edc7687cc9f04a38e118))
+
+
+### Features
+
+* adds direction property to design system ([#3535](https://github.com/Microsoft/fast/issues/3535)) ([492f5e7](https://github.com/Microsoft/fast/commit/492f5e766610a6cae51c9c90a18b83252a67f9d3))
+* basic menu item check state toggling ([#3512](https://github.com/Microsoft/fast/issues/3512)) ([ea84cbd](https://github.com/Microsoft/fast/commit/ea84cbd125b16de0ff2ae888f82163178af8da69))
+
+
+
+
+
 # [1.5.0](https://github.com/Microsoft/fast/compare/@microsoft/fast-components@1.4.0...@microsoft/fast-components@1.5.0) (2020-07-14)