+++ sites/fast-component-explorer/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.2](https://github.com/Microsoft/fast/compare/@microsoft/fast-component-explorer@0.11.1...@microsoft/fast-component-explorer@0.11.2) (2020-07-31)
+
+**Note:** Version bump only for package @microsoft/fast-component-explorer
+
+
+
+
+
+## [0.11.1](https://github.com/Microsoft/fast/compare/@microsoft/fast-component-explorer@0.11.0...@microsoft/fast-component-explorer@0.11.1) (2020-07-23)
+
+**Note:** Version bump only for package @microsoft/fast-component-explorer
+
+
+
+
+
 # [0.11.0](https://github.com/Microsoft/fast/compare/@microsoft/fast-component-explorer@0.10.0...@microsoft/fast-component-explorer@0.11.0) (2020-07-14)